refactor(api): split recurring-payments service into domain modules
Pure move of the 4000-line service into create/activation/collection/
lifecycle/update modules plus shared subscription-program plumbing;
public imports are unchanged via index re-exports.
