+++ apps/sdp-api/src/services/payments/recurring-payments/create.ts
@@ -1,0 +1,75 @@
+import {
+  createPaymentRecurringPaymentsRepository,
+  createPaymentsRepository,
+  type PaymentRecurringPaymentRow,
+} from "@/db/repositories";
+import { AppError } from "@/lib/errors";
+import type { CustodyWallet } from "@/services/stores/custody-config.store";
+import type { Env } from "@/types/env";
+import { resolveSolanaCounterpartyAccount } from "../counterparty-account-resolution";
+import { assertWalletPolicyAllowsTransferWithRepository } from "../wallet-policy";
+import { assertRecurringPaymentTokenMint } from "./shared";
+
+export async function createRecurringPayment(input: {
+  env: Env;
+  organizationId: string;
+  projectId: string;
+  sourceWallet: CustodyWallet;
+  counterpartyId: string;
+  counterpartyAccountId: string;
+  token: string;
+  amount: string;
+  periodHours: number;
+  firstCollectionAt?: string | null;
+  metadataUri?: string | null;
+  createdBy: string | null;
+}): Promise<PaymentRecurringPaymentRow> {
+  const [tokenMint, destination] = await Promise.all([
+    assertRecurringPaymentTokenMint(input.token, input.projectId, input.env),
+    resolveSolanaCounterpartyAccount({
+      env: input.env,
+      organizationId: input.organizationId,
+      projectId: input.projectId,
+      counterpartyId: input.counterpartyId,
+      counterpartyAccountId: input.counterpartyAccountId,
+    }),
+  ]);
+
+  await assertWalletPolicyAllowsTransferWithRepository(createPaymentsRepository(input.env), {
+    organizationId: input.organizationId,
+    projectId: input.projectId,
+    wallet: input.sourceWallet,
+    destinationAddress: destination.destinationAddress,
+    enforceDailyLimit: false,
+    token: tokenMint,
+    amount: input.amount,
+  });
+
+  const now = new Date().toISOString();
+  const recurringPayment = await createPaymentRecurringPaymentsRepository(
+    input.env
+  ).createRecurringPayment({
+    id: `prp_${crypto.randomUUID()}`,
+    organizationId: input.organizationId,
+    projectId: input.projectId,
+    sourceWalletId: input.sourceWallet.walletId,
+    sourceAddress: input.sourceWallet.publicKey,
+    counterpartyId: input.counterpartyId,
+    counterpartyAccountId: input.counterpartyAccountId,
+    destinationAddress: destination.destinationAddress,
+    token: tokenMint,
+    amount: input.amount,
+    periodHours: input.periodHours,
+    firstCollectionAt: input.firstCollectionAt ?? null,
+    metadataUri: input.metadataUri ?? null,
+    createdBy: input.createdBy,
+    createdAt: now,
+    updatedAt: now,
+  });
+
+  if (!recurringPayment) {
+    throw new AppError("INTERNAL_ERROR", "Failed to create recurring payment");
+  }
+
+  return recurringPayment;
+}
