+++ apps/sdp-api/src/services/payments/recurring-payments/index.ts
@@ -1,0 +1,5 @@
+export { activateRecurringPayment } from "./activation";
+export { collectRecurringPayment } from "./collection";
+export { createRecurringPayment } from "./create";
+export { cancelRecurringPayment, resumeRecurringPayment } from "./lifecycle";
+export { updateRecurringPayment } from "./update";
