+++ apps/sdp-api/src/services/payments/recurring-payments/shared.ts
@@ -1,0 +1,131 @@
+import { createFeePaymentAdapter } from "@sdp/payments/fee-payment";
+import { getSolanaConfig } from "@sdp/rpc";
+import * as solanaRpc from "@sdp/rpc/solana";
+import { assertValidAddress } from "@sdp/solana/address";
+import { WELL_KNOWN_TOKEN_BY_MINT } from "@sdp/types";
+import {
+  type Address,
+  addSignersToTransactionMessage,
+  appendTransactionMessageInstructions,
+  createTransactionMessage,
+  getTransactionEncoder,
+  type Instruction,
+  pipe,
+  type Signature,
+  setTransactionMessageFeePayer,
+  setTransactionMessageLifetimeUsingBlockhash,
+  type TransactionSigner,
+} from "@solana/kit";
+import { partiallySignTransactionMessageWithSigners } from "@solana/signers";
+import { createTokenRepository } from "@/db/repositories";
+import { AppError, badRequest } from "@/lib/errors";
+import { isNativePaymentToken, normalizePaymentToken } from "@/services/payment-operation.service";
+import * as solanaServices from "@/services/solana";
+import type { CustodyWallet } from "@/services/stores/custody-config.store";
+import type { Env } from "@/types/env";
+
+const RECURRING_PAYMENT_TOKEN_ERROR =
+  "Recurring payments support USD stablecoins and tokens issued in this project; native SOL is not supported";
+
+/**
+ * Resolves a recurring-payment token to its mint, allowing only well-known USD
+ * stablecoins on the configured cluster or active tokens issued in the project.
+ */
+export async function assertRecurringPaymentTokenMint(
+  token: string,
+  projectId: string,
+  env: Env
+): Promise<string> {
+  if (isNativePaymentToken(token)) {
+    throw badRequest(RECURRING_PAYMENT_TOKEN_ERROR);
+  }
+
+  const mint = assertValidAddress(normalizePaymentToken(token, env), "token");
+  const wellKnown = WELL_KNOWN_TOKEN_BY_MINT.get(mint);
+  if (wellKnown) {
+    const cluster = getSolanaConfig(env).network;
+    if (wellKnown.isUsdStable && wellKnown.mints[cluster] === mint) {
+      return mint;
+    }
+    throw badRequest(RECURRING_PAYMENT_TOKEN_ERROR);
+  }
+
+  const issuedTokenStatus = await createTokenRepository(env).getStatusByMint(projectId, mint);
+  if (issuedTokenStatus !== "active") {
+    throw badRequest(RECURRING_PAYMENT_TOKEN_ERROR);
+  }
+
+  return mint;
+}
+
+export function generateProgramPlanId(): string {
+  const bytes = new Uint8Array(8);
+  let value = 0n;
+
+  while (value === 0n) {
+    crypto.getRandomValues(bytes);
+    value = 0n;
+    for (const byte of bytes) {
+      value = (value << 8n) | BigInt(byte);
+    }
+  }
+
+  return value.toString();
+}
+
+export async function sendSubscriptionInstructions(input: {
+  env: Env;
+  organizationId: string;
+  projectId: string;
+  sourceWallet: CustodyWallet;
+  sourceSigner?: TransactionSigner;
+  instructions: Instruction[];
+  feePayer?: Address;
+}): Promise<Signature> {
+  const signer =
+    input.sourceSigner ??
+    (await solanaServices.createOrgSigner(
+      input.env,
+      input.organizationId,
+      input.projectId,
+      input.sourceWallet.walletId
+    ));
+
+  if (signer.address !== input.sourceWallet.publicKey) {
+    throw badRequest("Resolved signing wallet does not match source wallet");
+  }
+
+  const rpc = solanaRpc.createRpc(input.env);
+  const { blockhash, lastValidBlockHeight } = await solanaRpc.getRecentBlockhash(rpc, "confirmed");
+  const feePayment = createFeePaymentAdapter(input.env);
+  const feePayer = input.feePayer ?? (await feePayment.getFeePayer());
+  const message = pipe(
+    createTransactionMessage({ version: 0 }),
+    (m) => setTransactionMessageFeePayer(feePayer, m),
+    (m) => setTransactionMessageLifetimeUsingBlockhash({ blockhash, lastValidBlockHeight }, m),
+    (m) => appendTransactionMessageInstructions(input.instructions, m),
+    (m) => addSignersToTransactionMessage([signer], m)
+  );
+  const partiallySigned = await partiallySignTransactionMessageWithSigners(message);
+  const txBytes = new Uint8Array(getTransactionEncoder().encode(partiallySigned));
+  return feePayment.signAndSend(txBytes);
+}
+
+export async function confirmSubscriptionSignature(
+  env: Env,
+  signature: Signature,
+  message = "Recurring payment activation failed on-chain"
+): Promise<void> {
+  const rpc = solanaRpc.createRpc(env);
+  const confirmation = await solanaRpc.confirmTransaction(rpc, signature, {
+    commitment: "confirmed",
+  });
+
+  if (confirmation.err) {
+    throw new AppError("TRANSACTION_FAILED", message);
+  }
+}
+
+export function activationErrorMessage(error: unknown): string {
+  return error instanceof Error ? error.message : String(error);
+}
